Refactor into actual Python package
