+++ src/flexeval/classes/Metric.py
@@ -8,14 +8,15 @@ from pathlib import Path
 
 import peewee as pw
 import pydantic
-from classes.BaseModel import BaseModel
-from classes.Dataset import Dataset
-from classes.EvalSetRun import EvalSetRun
-from classes.Message import Message
-# from classes.DatasetRow import DatasetRow
-from classes.Thread import Thread
-from classes.ToolCall import ToolCall
-from classes.Turn import Turn
+from flexeval.classes.BaseModel import BaseModel
+from flexeval.classes.Dataset import Dataset
+from flexeval.classes.EvalSetRun import EvalSetRun
+from flexeval.classes.Message import Message
+
+# from flexeval.classes.DatasetRow import DatasetRow
+from flexeval.classes.Thread import Thread
+from flexeval.classes.ToolCall import ToolCall
+from flexeval.classes.Turn import Turn
 from playhouse.shortcuts import model_to_dict
