+++ src/flexeval/classes/Thread.py
@@ -2,9 +2,9 @@ import copy
 import json
 
 import peewee as pw
-from classes.BaseModel import BaseModel
-from classes.Dataset import Dataset
-from classes.EvalSetRun import EvalSetRun
+from flexeval.classes.BaseModel import BaseModel
+from flexeval.classes.Dataset import Dataset
+from flexeval.classes.EvalSetRun import EvalSetRun
 
 
 class Thread(BaseModel):
