+++ src/flexeval/compute_metrics.py
@@ -4,10 +4,10 @@ import json
 import string
 from typing import ForwardRef, Union, get_args
 
-from classes.Message import Message
-from classes.Thread import Thread
-from classes.ToolCall import ToolCall
-from classes.Turn import Turn
+from flexeval.classes.Message import Message
+from flexeval.classes.Thread import Thread
+from flexeval.classes.ToolCall import ToolCall
+from flexeval.classes.Turn import Turn
 
 from configuration import completion_functions, function_metrics
